Finalised architecture
The architecture is now fully functional. More work to be done in the future.
