+++ Checks/publicMutexMembers.py
@@ -1,13 +1,13 @@
-from Core.formalCheckInterface import *
-from typing import List
+import clang.cindex
+from alerts import Alert
 
 """
 TODO Write Description
 """
 
-class Check(FormalCheckInterface):
-    def analyse_cursor(self, cursor: clang.cindex.Cursor) -> List[Alert]:
-        alert_list: List[Alert] = list()
+class Check():
+    def analyse_cursor(self, cursor: clang.cindex.Cursor):
+        alert_list = list()
         
         if str(cursor.access_specifier) == "AccessSpecifier.PUBLIC":
                 count = 0
