--- src/colours.py
@@ -1,38 +1,0 @@
-# A list of ANSI colour codes for terminals. 'native' = default colours
-text_colours = {
-    "black"        : '\033[30m',
-    "red"          : '\033[31m',
-    "green"        : '\033[32m',
-    "yellow"       : '\033[33m',
-    "blue"         : '\033[34m',
-    "purple"       : '\033[35m',
-    "cyan"         : '\033[36m',
-    "light grey"   : '\033[37m',
-    "dark grey"    : '\033[90m',
-    "light red"    : '\033[91m',
-    "light green"  : '\033[92m',
-    "light yellow" : '\033[93m',
-    "light blue"   : '\033[94m',
-    "light purple" : '\033[95m',
-    "light cyan"   : '\033[96m',
-    "white"        : '\033[97m',
-}
-background_colours = {
-    "black"        : '\033[40m',
-    "red"          : '\033[41m',
-    "green"        : '\033[42m',
-    "yellow"       : '\033[43m',
-    "blue"         : '\033[44m',
-    "purple"       : '\033[45m',
-    "cyan"         : '\033[46m',
-    "light grey"   : '\033[47m',
-    "dark grey"    : '\033[100m',
-    "light red"    : '\033[101m',
-    "light green"  : '\033[102m',
-    "light yellow" : '\033[103m',
-    "light blue"   : '\033[104m',
-    "light purple" : '\033[105m',
-    "light cyan"   : '\033[106m',
-    "white"        : '\033[107m',
-    "native"       : '\033[m'
-}
